backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
anthropic==0.15.0
python-dotenv==1.0.0 